+++ templates/admin/index_m.html.ts
@@ -36,20 +36,22 @@ export = (answers => {
 
 	<!-- Load ioBroker scripts and styles-->
 	<link rel="stylesheet" type="text/css" href="../../css/adapter.css" />
-	<link rel="stylesheet" type="text/css" href="../../lib/css/materialize.css">
+${useReact ? 
+`	<script type="text/javascript" src="../../socket.io/socket.io.js"></script>` :
+`	<link rel="stylesheet" type="text/css" href="../../lib/css/materialize.css">
 
 	<script type="text/javascript" src="../../lib/js/jquery-3.2.1.min.js"></script>
 	<script type="text/javascript" src="../../socket.io/socket.io.js"></script>
 
 	<script type="text/javascript" src="../../js/translate.js"></script>
 	<script type="text/javascript" src="../../lib/js/materialize.js"></script>
-	<script type="text/javascript" src="../../js/adapter-settings.js"></script>
+	<script type="text/javascript" src="../../js/adapter-settings.js"></script>`}
 
 	<!-- Load our own files -->
 	<link rel="stylesheet" type="text/css" href="style.css" />
-	<script type="text/javascript" src="words.js"></script>
+${useReact ? "" :
+`	<script type="text/javascript" src="words.js"></script>
 
-${useReact ? "" : (`
 	<script type="text/javascript">
 		// This will be called by the admin adapter when the settings page loads
 		function load(settings, onChange) {
@@ -91,19 +93,14 @@ ${useReact ? "" : (`
 			callback(obj);
 		}
 	</script>
-`)}
+`}
 </head>
 
 <body>
 
 ${useReact ? (`
 	<!-- this is where the React components are loaded into -->
-	<div class="m adapter-container">
-		<div class="loading">
-			<img src="${answers.adapterName}.png" />
-		</div>
-		<div class="translate loading">loading...</div>
-	</div>
+	<div class="m adapter-container" id="root"></div>
 
 	<!-- load compiled React scripts -->
 	<script type="text/javascript" src="build/index.js"></script>