+++ templates/lib/adapter-config.d.ts.ts
@@ -1,4 +1,21 @@
 import { TemplateFunction } from "../../src/lib/createAdapter";
+import { AdapterSettings, getDefaultAnswer } from "../../src/lib/questions";
+
+function generateSettingsProperty(settings: AdapterSettings): string {
+	if (settings.inputType === "select" && settings.options) {
+		return `
+			${settings.key}: (${settings.options.map((opt) => `"${opt.value}"`).join(" | ")});`;
+	} else if (settings.inputType === "checkbox") {
+		return `
+			${settings.key}: boolean;`;
+	} else if (settings.inputType === "number") {
+		return `
+			${settings.key}: number;`;
+	} else {
+		return `
+			${settings.key}: string;`;
+	}
+}
 
 const templateFunction: TemplateFunction = answers => {
 
@@ -7,9 +24,26 @@ const templateFunction: TemplateFunction = answers => {
 
 	const useTypeScript = answers.language === "TypeScript";
 	const useTypeChecking = answers.tools && answers.tools.indexOf("type checking") > -1;
-	if (useTypeScript && !useTypeChecking) return; // Don't do the copy/delete stuff in the first version... We'll add this later
+	let template: string;
+	if (useTypeScript && !useTypeChecking) {
+		const adapterSettings: AdapterSettings[] = answers.adapterSettings ?? getDefaultAnswer("adapterSettings")!;
+		
+		template = `
+// This file extends the AdapterConfig type from "@types/iobroker"
 
-	const template = `
+// Augment the globally declared type ioBroker.AdapterConfig
+declare global {
+	namespace ioBroker {
+		interface AdapterConfig {${adapterSettings.map(generateSettingsProperty).join("")}
+		}
+	}
+}
+
+// this is required so the above AdapterConfig is found by TypeScript / type checking
+export {};
+`;
+	} else {
+		template = `
 // This file extends the AdapterConfig type from "@types/iobroker"
 // using the actual properties present in io-package.json
 // in order to provide typings for adapter.config properties
@@ -26,7 +60,12 @@ declare global {
 		}
 	}
 }
+
+// this is required so the above AdapterConfig is found by TypeScript / type checking
+export {};
 `;
+	}
+
 	return template.trim();
 };