+++ templates/src/main.es6.ts
@@ -25,20 +25,6 @@ import * as utils from "@iobroker/adapter-core";
 // Load your modules here, e.g.:
 // import * as fs from "fs";
 
-// Augment the adapter.config object with the actual types
-// TODO: delete this in the next version
-declare global {
-	// eslint-disable-next-line @typescript-eslint/no-namespace
-	namespace ioBroker {
-		interface AdapterConfig {
-			// Define the shape of your options here (recommended)
-${adapterSettings.map(s => `\t\t\t${s.key}: ${typeof s.defaultValue};`).join("\n")}
-			// Or use a catch-all approach
-			[key: string]: any;
-		}
-	}
-}
-
 class ${className} extends utils.Adapter {
 
 	public constructor(options: Partial<utils.AdapterOptions> = {}) {