+++ test/baselines/adapter_JS_ES6Class_ESLint_TypeChecking_Spaces_SingleQuotes_Apache-2.0/lib/adapter-config.d.ts
@@ -14,3 +14,6 @@ declare global {
         }
     }
 }
+
+// this is required so the above AdapterConfig is found by TypeScript / type checking
+export {};